+++ chalicelib/processors/GenerateProcessor.py
@@ -39,8 +39,8 @@ class GenerateProcessor(GenericProcessor):
         if code is None:
             raise BadRequestError("Error: please provide the original code")
 
-        # The output language is optional; if not set, then default to Python
-        outputlanguage = data.get('language', 'python')
+        # The output language is optional; if not set, then default to existing language
+        outputlanguage = data.get('language', 'programming')
 
         result = self.process_input(data, account, function_name, correlation_id,
                                     {self.get_chunkable_input(): explanation,